@@ -12,14 +12,14 @@
     "@ui5/webcomponents": "1.24.0",
     "@ui5/webcomponents-fiori": "1.24.0",
     "@ui5/webcomponents-react": "1.27.0",
-    "@types/node": "20.12.5",
-    "@types/react": "18.2.74",
-    "@types/react-dom": "18.2.24",
+    "@types/node": "20.12.7",
+    "@types/react": "18.2.78",
+    "@types/react-dom": "18.2.25",
     "eslint": "8.57.0",
-    "eslint-config-next": "14.1.4",
-    "next": "14.1.4",
+    "eslint-config-next": "14.2.1",
+    "next": "14.2.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "typescript": "5.4.4"
+    "typescript": "5.4.5"
   }
 }
