@@ -1,25 +1,9 @@
 import type { Virtualizer } from '@tanstack/react-virtual';
-import { ThemingParameters } from '@ui5/webcomponents-react-base';
+import { useStylesheet } from '@ui5/webcomponents-react-base';
 import React, { forwardRef, Fragment } from 'react';
-import { createUseStyles } from 'react-jss';
 import type { DivWithCustomScrollProp } from '../types/index.js';
+import { classNames, styleData } from './Resizer.module.css.js';
 import { ColumnHeader } from './index.js';
-
-const styles = {
-  resizer: {
-    display: 'inline-block',
-    width: '3px',
-    height: '100%',
-    position: 'absolute',
-    bottom: 0,
-    top: 0,
-    zIndex: 1,
-    cursor: 'col-resize',
-    '&:hover, &:active': {
-      backgroundColor: ThemingParameters.sapContent_DragAndDropActiveColor
-    }
-  }
-};
 
 interface ColumnHeaderContainerProps {
   headerProps: Record<string, unknown>;
@@ -35,8 +19,6 @@ interface ColumnHeaderContainerProps {
   showVerticalEndBorder: boolean;
 }
 
-const useStyles = createUseStyles(styles, { name: 'Resizer' });
-
 export const ColumnHeaderContainer = forwardRef<HTMLDivElement, ColumnHeaderContainerProps>((props, ref) => {
   const {
     headerProps,
@@ -51,7 +33,7 @@ export const ColumnHeaderContainer = forwardRef<HTMLDivElement, ColumnHeaderCont
     showVerticalEndBorder
   } = props;
 
-  const classes = useStyles();
+  useStylesheet(styleData, 'Resizer');
 
   return (
     <div
@@ -83,7 +65,7 @@ export const ColumnHeaderContainer = forwardRef<HTMLDivElement, ColumnHeaderCont
               <div
                 {...column.getResizerProps()}
                 data-resizer
-                className={classes.resizer}
+                className={classNames.resizer}
                 style={resizerDirectionStyle}
               />
             )}
