@@ -3,7 +3,7 @@ import iconFilter from '@ui5/webcomponents-icons/dist/filter.js';
 import iconGroup from '@ui5/webcomponents-icons/dist/group-2.js';
 import iconSortAscending from '@ui5/webcomponents-icons/dist/sort-ascending.js';
 import iconSortDescending from '@ui5/webcomponents-icons/dist/sort-descending.js';
-import { ThemingParameters } from '@ui5/webcomponents-react-base';
+import { ThemingParameters, useStylesheet } from '@ui5/webcomponents-react-base';
 import { clsx } from 'clsx';
 import type {
   AriaAttributes,
@@ -14,12 +14,11 @@ import type {
   ReactNode
 } from 'react';
 import React, { useRef, useState } from 'react';
-import { createUseStyles } from 'react-jss';
-import { CustomThemingParameters } from '../../../themes/CustomVariables.js';
 import { Icon } from '../../../webComponents/Icon/index.js';
 import { Text } from '../../Text/index.js';
 import type { ColumnType } from '../types/ColumnType.js';
 import type { DivWithCustomScrollProp } from '../types/index.js';
+import { classNames, styleData } from './ColumnHeader.module.css.js';
 import { ColumnHeaderModal } from './ColumnHeaderModal.js';
 
 export interface ColumnHeaderProps {
@@ -56,66 +55,8 @@ export interface ColumnHeaderProps {
   ['aria-label']?: AriaAttributes['aria-label'];
 }
 
-const styles = {
-  thContainer: {
-    '&:first-child': {
-      '& > [role="columnheader"]': {
-        borderInlineStart: CustomThemingParameters.AnalyticalTableOuterCellBorder
-      }
-    },
-    '&:last-child': {
-      '& > [role="columnheader"]': {
-        borderInlineEnd: CustomThemingParameters.AnalyticalTableOuterCellBorder
-      }
-    }
-  },
-  verticalEndBorder: {
-    '&:last-child': {
-      '& > [role="columnheader"]': {
-        borderInlineEnd: `1px solid ${ThemingParameters.sapList_BorderColor}`
-      }
-    }
-  },
-  header: {
-    height: '100%',
-    display: 'flex',
-    alignItems: 'center',
-    textAlign: 'start',
-    fontFamily: CustomThemingParameters.AnalyticalTableHeaderFontFamily,
-    fontSize: ThemingParameters.sapFontSize,
-    fontWeight: 'normal',
-    color: 'inherit',
-    textOverflow: 'ellipsis',
-    whiteSpace: 'nowrap',
-    maxWidth: '100%',
-    position: 'relative',
-    width: '100%',
-    overflowX: 'hidden',
-    overflowY: 'hidden',
-    boxSizing: 'border-box'
-  },
-  text: {
-    color: 'inherit',
-    fontFamily: 'inherit',
-    width: '100%',
-    textAlign: 'start'
-  },
-  iconContainer: {
-    display: 'flex',
-    position: 'absolute',
-    color: ThemingParameters.sapContent_IconColor,
-    insetInlineEnd: '0.5rem'
-  },
-  selectAllCheckBoxContainer: {
-    display: 'flex',
-    justifyContent: 'center'
-  }
-};
-
-const useStyles = createUseStyles(styles, { name: 'TableColumnHeader' });
-
 export const ColumnHeader = (props: ColumnHeaderProps) => {
-  const classes = useStyles();
+  useStylesheet(styleData, ColumnHeader.displayName);
   const {
     id,
     children,
@@ -219,7 +160,7 @@ export const ColumnHeader = (props: ColumnHeaderProps) => {
   return (
     <div
       ref={columnHeaderRef}
-      className={clsx(classes.thContainer, showVerticalEndBorder && classes.verticalEndBorder)}
+      className={clsx(classNames.thContainer, showVerticalEndBorder && classNames.verticalEndBorder)}
       style={{
         position: 'absolute',
         top: 0,
@@ -256,21 +197,21 @@ export const ColumnHeader = (props: ColumnHeaderProps) => {
         aria-sort={ariaSort}
         title={title}
       >
-        <div className={classes.header} data-h-align={column.hAlign}>
+        <div className={classNames.header} data-h-align={column.hAlign}>
           <Text
             title={tooltip}
             wrapping={false}
             style={textStyle}
             className={clsx(
-              classes.text,
-              columnId === '__ui5wcr__internal_selection_column' && classes.selectAllCheckBoxContainer
+              classNames.text,
+              columnId === '__ui5wcr__internal_selection_column' && classNames.selectAllCheckBoxContainer
             )}
             data-component-name={`AnalyticalTableHeaderHeaderContentContainer-${columnId}`}
           >
             {children}
           </Text>
           <div
-            className={classes.iconContainer}
+            className={classNames.iconContainer}
             data-component-name={`AnalyticalTableHeaderIconsContainer-${columnId}`}
           >
             {isFiltered && <Icon name={iconFilter} aria-hidden />}
@@ -296,3 +237,5 @@ export const ColumnHeader = (props: ColumnHeaderProps) => {
     </div>
   );
 };
+
+ColumnHeader.displayName = 'ColumnHeader';
