@@ -1,34 +1,17 @@
-import { ThemingParameters } from '@ui5/webcomponents-react-base';
+import { useStylesheet } from '@ui5/webcomponents-react-base';
 import React from 'react';
-import { createUseStyles } from 'react-jss';
 import { makeRenderer } from 'react-table';
 import { FlexBoxAlignItems, FlexBoxDirection, FlexBoxWrap } from '../../../../enums/index.js';
 import { FlexBox } from '../../../FlexBox/index.js';
 import { Text } from '../../../Text/index.js';
-
-const PopInStyles = {
-  container: {
-    width: '100%'
-  },
-  defaultCell: {
-    fontFamily: `${ThemingParameters.sapFontBoldFamily}`,
-    '& *': {
-      fontFamily: `${ThemingParameters.sapFontBoldFamily}`
-    }
-  },
-  header: {
-    height: '16px'
-  }
-};
-
-const useStyles = createUseStyles(PopInStyles, { name: 'PopIn' });
+import { classNames, styleData } from './PopIn.module.css.js';
 
 export const PopIn = (instance) => {
   const { state, contentToRender, cell, row, internalRowHeight } = instance;
 
-  const classes = useStyles();
+  useStylesheet(styleData, PopIn.displayName);
   return (
-    <FlexBox direction={FlexBoxDirection.Column} className={classes.container}>
+    <FlexBox direction={FlexBoxDirection.Column} className={classNames.container}>
       <FlexBox
         alignItems={
           contentToRender !== 'Grouped' && contentToRender !== 'Expandable'
@@ -36,7 +19,7 @@ export const PopIn = (instance) => {
             : FlexBoxAlignItems.Center
         }
         wrap={FlexBoxWrap.NoWrap}
-        className={classes.defaultCell}
+        className={classNames.defaultCell}
         style={{
           height: internalRowHeight
         }}
@@ -76,7 +59,7 @@ export const PopIn = (instance) => {
           };
           return (
             <FlexBox direction={FlexBoxDirection.Column} key={item.id}>
-              {item.column?.Header && <div className={classes.header}>{renderHeader()}:</div>}
+              {item.column?.Header && <div className={classNames.header}>{renderHeader()}:</div>}
               <div style={{ height: internalRowHeight }}>{popInInstanceProps && renderCell()}</div>
             </FlexBox>
           );
@@ -84,3 +67,5 @@ export const PopIn = (instance) => {
     </FlexBox>
   );
 };
+
+PopIn.displayName = 'PopIn';
