@@ -1,22 +1,11 @@
-import { CssSizeVariables } from '@ui5/webcomponents-react-base';
+import { useStylesheet } from '@ui5/webcomponents-react-base';
 import React from 'react';
-import { createUseStyles } from 'react-jss';
 import { Loader } from '../../../Loader/index.js';
-
-const styles = {
-  loader: {
-    position: 'absolute',
-    top: CssSizeVariables.ui5WcrAnalyticalTableHeaderRowHeight,
-    left: 0,
-    right: 0,
-    zIndex: 1
-  }
-};
-const useStyles = createUseStyles(styles, { name: 'LoadingComponent' });
+import { classNames, styleData } from './DefaultLoadingComponent.module.css.js';
 
 const DefaultLoadingComponent = ({ style }) => {
-  const classes = useStyles();
-  return <Loader delay={500} style={style} className={classes.loader} />;
+  useStylesheet(styleData, DefaultLoadingComponent.displayName);
+  return <Loader delay={500} style={style} className={classNames.loader} />;
 };
 
 DefaultLoadingComponent.displayName = 'DefaultLoadingComponent';
