@@ -4,16 +4,16 @@ import iconPushPinOff from '@ui5/webcomponents-icons/dist/pushpin-off.js';
 import iconPushPinOn from '@ui5/webcomponents-icons/dist/pushpin-on.js';
 import iconArrowDown from '@ui5/webcomponents-icons/dist/slim-arrow-down.js';
 import iconArrowUp from '@ui5/webcomponents-icons/dist/slim-arrow-up.js';
-import { enrichEventWithDetails, ThemingParameters, useI18nBundle } from '@ui5/webcomponents-react-base';
+import { enrichEventWithDetails, useI18nBundle, useStylesheet } from '@ui5/webcomponents-react-base';
 import { clsx } from 'clsx';
-import React, { forwardRef, useCallback, useEffect, useRef } from 'react';
 import type { CSSProperties } from 'react';
-import { createUseStyles } from 'react-jss';
+import React, { forwardRef, useCallback, useEffect, useRef } from 'react';
 import { COLLAPSE_HEADER, EXPAND_HEADER, PIN_HEADER, UNPIN_HEADER } from '../../i18n/i18n-defaults.js';
 import { cssVarVersionInfoPrefix, getUi5TagWithSuffix } from '../../internal/utils.js';
 import type { CommonProps } from '../../types/index.js';
 import { Button, ToggleButton } from '../../webComponents/index.js';
 import type { ButtonDomRef } from '../../webComponents/index.js';
+import { classNames, styleData } from './DynamicPageAnchorBar.module.css.js';
 
 const _buttonBaseMinWidth = `${cssVarVersionInfoPrefix}button_base_min_width`;
 const _buttonBaseHeight = `${cssVarVersionInfoPrefix}button_base_height`;
@@ -22,51 +22,6 @@ const anchorButtonVariables = {
   [_buttonBaseMinWidth]: '1.5rem',
   [_buttonBaseHeight]: '1.5rem'
 } as CSSProperties;
-
-const anchorBarStyles = {
-  container: {
-    position: 'absolute',
-    left: '50%',
-    '&:before, &:after': {
-      content: '""',
-      position: 'absolute',
-      width: '4rem',
-      top: '50%',
-      height: '0.0625rem'
-    },
-    '&:before': {
-      right: '100%',
-      backgroundImage: `linear-gradient(to left, ${ThemingParameters.sapObjectHeader_BorderColor}, transparent)`
-    },
-    '&:after': {
-      backgroundImage: `linear-gradient(to right, ${ThemingParameters.sapObjectHeader_BorderColor}, transparent)`,
-      left: '100%'
-    }
-  },
-  anchorBarActionButton: {
-    '--_ui5wcr_anchor-btn-center': `calc((var(${_buttonBaseMinWidth}) - ${ThemingParameters.sapButton_BorderWidth}) / 2)`,
-    position: 'absolute',
-    insetBlockStart: `calc(-1 * var(--_ui5wcr_anchor-btn-center))`,
-    insetInlineStart: 'calc(50% - var(--_ui5wcr_anchor-btn-center))',
-    zIndex: 3,
-    '&:not([pressed])': {
-      backgroundColor: ThemingParameters.sapObjectHeader_Background
-    }
-  },
-  anchorBarActionButtonExpandable: {
-    '& + $anchorBarActionButtonPinnable': {
-      insetInlineStart: 'calc(50% - var(--_ui5wcr_anchor-btn-center) + 1rem)'
-    }
-  },
-  anchorBarActionButtonPinnable: {},
-  anchorBarActionPinnableAndExpandable: {
-    '&$anchorBarActionButtonExpandable': {
-      insetInlineStart: 'calc(50% - var(--_ui5wcr_anchor-btn-center) - 1rem)'
-    }
-  }
-};
-
-const useStyles = createUseStyles(anchorBarStyles, { name: 'DynamicPageAnchorBar' });
 
 interface DynamicPageAnchorBarPropTypes extends CommonProps {
   /**
@@ -129,8 +84,8 @@ const DynamicPageAnchorBar = forwardRef<HTMLElement, DynamicPageAnchorBarPropTyp
     onHoverToggleButton
   } = props;
 
+  useStylesheet(styleData, DynamicPageAnchorBar.displayName);
   const showHideHeaderBtnRef = useRef<ButtonDomRef>(null);
-  const classes = useStyles();
   const shouldRenderHeaderPinnableButton = headerContentPinnable && headerContentVisible;
   const showBothActions = shouldRenderHeaderPinnableButton && showHideHeaderButton;
 
@@ -170,7 +125,7 @@ const DynamicPageAnchorBar = forwardRef<HTMLElement, DynamicPageAnchorBarPropTyp
       data-component-name="DynamicPageAnchorBar"
       style={style}
       role={a11yConfig?.dynamicPageAnchorBar?.role}
-      className={showHideHeaderButton || headerContentPinnable ? classes.container : null}
+      className={showHideHeaderButton || headerContentPinnable ? classNames.container : null}
       ref={ref}
     >
       {showHideHeaderButton && (
@@ -179,9 +134,9 @@ const DynamicPageAnchorBar = forwardRef<HTMLElement, DynamicPageAnchorBarPropTyp
           icon={!headerContentVisible ? iconArrowDown : iconArrowUp}
           data-ui5wcr-dynamic-page-header-action=""
           className={clsx(
-            classes.anchorBarActionButton,
-            classes.anchorBarActionButtonExpandable,
-            showBothActions && classes.anchorBarActionPinnableAndExpandable
+            classNames.anchorBarActionButton,
+            classNames.anchorBarActionButtonExpandable,
+            showBothActions && classNames.anchorBarActionPinnableAndExpandable
           )}
           style={anchorButtonVariables}
           onClick={onToggleHeaderButtonClick}
@@ -196,7 +151,7 @@ const DynamicPageAnchorBar = forwardRef<HTMLElement, DynamicPageAnchorBarPropTyp
         <ToggleButton
           icon={headerPinned ? iconPushPinOn : iconPushPinOff}
           data-ui5wcr-dynamic-page-header-action=""
-          className={clsx(classes.anchorBarActionButton, classes.anchorBarActionButtonPinnable)}
+          className={clsx(classNames.anchorBarActionButton, classNames.anchorBarActionButtonPinnable)}
           style={anchorButtonVariables}
           pressed={headerPinned}
           onClick={onPinHeader}
