@@ -1,22 +1,19 @@
 'use client';
 
-import { Device, useSyncRef } from '@ui5/webcomponents-react-base';
+import { Device, useStylesheet, useSyncRef } from '@ui5/webcomponents-react-base';
 import { clsx } from 'clsx';
 import type { ElementType, ReactNode } from 'react';
 import React, { forwardRef, useCallback, useEffect, useMemo, useReducer, useRef, useState } from 'react';
-import { createUseStyles } from 'react-jss';
 import { FormBackgroundDesign, TitleLevel } from '../../enums/index.js';
 import type { CommonProps } from '../../types/index.js';
 import { Title } from '../../webComponents/index.js';
-import { styles } from './Form.jss.js';
+import { classNames, styleData } from './Form.module.css.js';
 import { FormContext } from './FormContext.js';
 import type { FormContextType, FormElementTypes, FormGroupLayoutInfo, FormItemLayoutInfo, ItemInfo } from './types.js';
 
 const recalcReducerFn = (prev: number) => {
   return prev + 1;
 };
-
-const useStyles = createUseStyles(styles, { name: 'Form' });
 
 export interface FormPropTypes extends CommonProps {
   /**
@@ -130,7 +127,7 @@ const Form = forwardRef<HTMLFormElement, FormPropTypes>((props, ref) => {
   } = props;
 
   const [items, setItems] = useState<Map<string, ItemInfo>>(() => new Map());
-  const classes = useStyles();
+  useStylesheet(styleData, Form.displayName);
 
   const columnsMap = new Map();
   columnsMap.set('Phone', columnsS);
@@ -265,7 +262,7 @@ const Form = forwardRef<HTMLFormElement, FormPropTypes>((props, ref) => {
 
     return { formItems, formGroups, registerItem, unregisterItem, rowsWithGroup };
   }, [items, registerItem, unregisterItem, currentNumberOfColumns, titleText, currentLabelSpan]);
-  const formClassNames = clsx(classes.form, classes[backgroundDesign.toLowerCase()]);
+  const formClassNames = clsx(classNames.form, classNames[backgroundDesign.toLowerCase()]);
   const CustomTag = as as ElementType;
 
   const prevFormItems = useRef<undefined | FormItemLayoutInfo[]>(undefined);
@@ -298,7 +295,7 @@ const Form = forwardRef<HTMLFormElement, FormPropTypes>((props, ref) => {
   return (
     <FormContext.Provider value={{ ...formLayoutContextValue, labelSpan: currentLabelSpan, recalcTrigger }}>
       <CustomTag
-        className={clsx(classes.formContainer, className)}
+        className={clsx(classNames.formContainer, className)}
         suppressHydrationWarning={true}
         ref={componentRef}
         style={{
@@ -316,7 +313,7 @@ const Form = forwardRef<HTMLFormElement, FormPropTypes>((props, ref) => {
       >
         <div className={formClassNames}>
           {titleText && (
-            <Title level={TitleLevel.H3} className={classes.formTitle} style={{ gridColumn: '1 / -1' }}>
+            <Title level={TitleLevel.H3} className={classNames.formTitle} style={{ gridColumn: '1 / -1' }}>
               {titleText}
             </Title>
           )}
