@@ -6,6 +6,7 @@ import {
   deprecationNotice,
   enrichEventWithDetails,
   ThemingParameters,
+  useStylesheet,
   useSyncRef
 } from '@ui5/webcomponents-react-base';
 import { clsx } from 'clsx';
@@ -20,7 +21,6 @@ import React, {
   useRef,
   useState
 } from 'react';
-import { createUseStyles } from 'react-jss';
 import { AvatarSize, GlobalStyleClasses, ObjectPageMode } from '../../enums/index.js';
 import { addCustomCSSWithScoping } from '../../internal/addCustomCSSWithScoping.js';
 import { safeGetChildrenArray } from '../../internal/safeGetChildrenArray.js';
@@ -28,12 +28,12 @@ import { useObserveHeights } from '../../internal/useObserveHeights.js';
 import type { CommonProps, Ui5CustomEvent } from '../../types/index.js';
 import type { AvatarPropTypes, TabContainerDomRef } from '../../webComponents/index.js';
 import { Tab, TabContainer } from '../../webComponents/index.js';
-import { DynamicPageCssVariables } from '../DynamicPage/DynamicPage.jss.js';
+import { DynamicPageCssVariables } from '../DynamicPage/utils.js';
 import { DynamicPageAnchorBar } from '../DynamicPageAnchorBar/index.js';
 import { DynamicPageHeader } from '../DynamicPageHeader/index.js';
 import type { ObjectPageSectionPropTypes } from '../ObjectPageSection/index.js';
 import { CollapsedAvatar } from './CollapsedAvatar.js';
-import { styles } from './ObjectPage.jss.js';
+import { classNames, styleData } from './ObjectPage.module.css.js';
 import { extractSectionIdFromHtmlId, getSectionById } from './ObjectPageUtils.js';
 
 addCustomCSSWithScoping(
@@ -173,8 +173,6 @@ export interface ObjectPagePropTypes extends Omit<CommonProps, 'placeholder'> {
   onPinnedStateChange?: (pinned: boolean) => void;
 }
 
-const useStyles = createUseStyles(styles, { name: 'ObjectPage' });
-
 /**
  * A component that allows apps to easily display information related to a business object.
  *
@@ -206,7 +204,7 @@ const ObjectPage = forwardRef<HTMLDivElement, ObjectPagePropTypes>((props, ref) 
     ...rest
   } = props;
 
-  const classes = useStyles();
+  useStylesheet(styleData, ObjectPage.displayName);
 
   const firstSectionId: string | undefined = safeGetChildrenArray<ReactElement>(children)[0]?.props?.id;
 
@@ -304,19 +302,19 @@ const ObjectPage = forwardRef<HTMLDivElement, ObjectPagePropTypes>((props, ref) 
     if (typeof image === 'string') {
       return (
         <span
-          className={classes.headerImage}
+          className={classNames.headerImage}
           style={{ borderRadius: imageShapeCircle ? '50%' : 0, overflow: 'hidden' }}
         >
-          <img src={image} className={classes.image} alt="Company Logo" />
+          <img src={image} className={classNames.image} alt="Company Logo" />
         </span>
       );
     } else {
       return cloneElement(image, {
         size: AvatarSize.L,
-        className: clsx(classes.headerImage, image.props?.className)
+        className: clsx(classNames.headerImage, image.props?.className)
       } as AvatarPropTypes);
     }
-  }, [image, classes.headerImage, classes.image, imageShapeCircle]);
+  }, [image, classNames.headerImage, classNames.image, imageShapeCircle]);
 
   const scrollToSectionById = (id?: string, isSubSection = false) => {
     const section = objectPageRef.current?.querySelector<HTMLElement>(
@@ -500,11 +498,11 @@ const ObjectPage = forwardRef<HTMLDivElement, ObjectPagePropTypes>((props, ref) 
     scrollTimeout.current = performance.now() + 500;
     if (!e.detail.visible) {
       setHeaderCollapsedInternal(true);
-      objectPageRef.current?.classList.add(classes.headerCollapsed);
+      objectPageRef.current?.classList.add(classNames.headerCollapsed);
     } else {
       setHeaderCollapsedInternal(false);
       setScrolledHeaderExpanded(true);
-      objectPageRef.current?.classList.remove(classes.headerCollapsed);
+      objectPageRef.current?.classList.remove(classNames.headerCollapsed);
     }
   }, []);
 
@@ -530,10 +528,10 @@ const ObjectPage = forwardRef<HTMLDivElement, ObjectPagePropTypes>((props, ref) 
   );
 
   const objectPageClasses = clsx(
-    classes.objectPage,
+    classNames.objectPage,
     GlobalStyleClasses.sapScrollBar,
     className,
-    mode === ObjectPageMode.IconTabBar && classes.iconTabBarMode
+    mode === ObjectPageMode.IconTabBar && classNames.iconTabBarMode
   );
 
   const { onScroll: _0, selectedSubSectionId: _1, ...propsWithoutOmitted } = rest;
@@ -626,7 +624,7 @@ const ObjectPage = forwardRef<HTMLDivElement, ObjectPagePropTypes>((props, ref) 
   const hasHeaderContent = !!headerContent;
   const renderTitleSection = useCallback(
     (inHeader = false) => {
-      const titleInHeaderClass = inHeader ? classes.titleInHeader : undefined;
+      const titleInHeaderClass = inHeader ? classNames.titleInHeader : undefined;
 
       if (headerTitle?.props && headerTitle.props?.showSubHeaderRight === undefined) {
         return cloneElement(headerTitle, {
@@ -670,7 +668,7 @@ const ObjectPage = forwardRef<HTMLDivElement, ObjectPagePropTypes>((props, ref) 
         headerPinned: headerPinned || scrolledHeaderExpanded,
         ref: componentRefHeaderContent,
         children: (
-          <div className={classes.headerContainer} data-component-name="ObjectPageHeaderContainer">
+          <div className={classNames.headerContainer} data-component-name="ObjectPageHeaderContainer">
             {avatar}
             {(headerContent.props.children || titleInHeader) && (
               <div data-component-name="ObjectPageHeaderContent">
@@ -689,7 +687,7 @@ const ObjectPage = forwardRef<HTMLDivElement, ObjectPagePropTypes>((props, ref) 
           headerPinned={headerPinned || scrolledHeaderExpanded}
           ref={componentRefHeaderContent}
         >
-          <div className={classes.headerContainer} data-component-name="ObjectPageHeaderContainer">
+          <div className={classNames.headerContainer} data-component-name="ObjectPageHeaderContainer">
             {avatar}
             <div data-component-name="ObjectPageHeaderContent">{titleInHeader && renderTitleSection(true)}</div>
           </div>
@@ -759,7 +757,7 @@ const ObjectPage = forwardRef<HTMLDivElement, ObjectPagePropTypes>((props, ref) 
         setIsAfterScroll(true);
       }, 100);
       if (!headerPinned || e.target.scrollTop === 0) {
-        objectPageRef.current?.classList.remove(classes.headerCollapsed);
+        objectPageRef.current?.classList.remove(classNames.headerCollapsed);
       }
       if (scrolledHeaderExpanded && e.target.scrollTop !== prevScrollTop.current) {
         if (e.target.scrollHeight - e.target.scrollTop === e.target.clientHeight) {
@@ -778,12 +776,12 @@ const ObjectPage = forwardRef<HTMLDivElement, ObjectPagePropTypes>((props, ref) 
   const onHoverToggleButton = useCallback(
     (e) => {
       if (e?.type === 'mouseover') {
-        topHeaderRef.current?.classList.add(classes.headerHoverStyles);
+        topHeaderRef.current?.classList.add(classNames.headerHoverStyles);
       } else {
-        topHeaderRef.current?.classList.remove(classes.headerHoverStyles);
+        topHeaderRef.current?.classList.remove(classNames.headerHoverStyles);
       }
     },
-    [classes.headerHoverStyles]
+    [classNames.headerHoverStyles]
   );
 
   const objectPageStyles: CSSProperties = {
@@ -811,7 +809,7 @@ const ObjectPage = forwardRef<HTMLDivElement, ObjectPagePropTypes>((props, ref) 
         role={a11yConfig?.objectPageTopHeader?.role}
         data-not-clickable={titleHeaderNotClickable}
         aria-roledescription={a11yConfig?.objectPageTopHeader?.ariaRoledescription ?? 'Object Page header'}
-        className={classes.header}
+        className={classNames.header}
         onClick={onTitleClick}
         style={{
           gridAutoColumns: `min-content ${
@@ -825,7 +823,7 @@ const ObjectPage = forwardRef<HTMLDivElement, ObjectPagePropTypes>((props, ref) 
         )}
         {headerTitle && renderTitleSection()}
         {snappedHeaderInObjPage && (
-          <div className={classes.snappedContent} data-component-name="ATwithImageSnappedContentContainer">
+          <div className={classNames.snappedContent} data-component-name="ATwithImageSnappedContentContainer">
             {headerTitle.props.snappedContent}
           </div>
         )}
@@ -835,7 +833,7 @@ const ObjectPage = forwardRef<HTMLDivElement, ObjectPagePropTypes>((props, ref) 
         <div
           data-component-name="ObjectPageAnchorBar"
           ref={anchorBarRef}
-          className={classes.anchorBar}
+          className={classNames.anchorBar}
           style={{
             top:
               scrolledHeaderExpanded || headerPinned
@@ -859,7 +857,7 @@ const ObjectPage = forwardRef<HTMLDivElement, ObjectPagePropTypes>((props, ref) 
       {!placeholder && (
         <div
           ref={tabContainerContainerRef}
-          className={classes.tabContainer}
+          className={classNames.tabContainer}
           data-component-name="ObjectPageTabContainer"
           style={{
             top:
@@ -873,7 +871,7 @@ const ObjectPage = forwardRef<HTMLDivElement, ObjectPagePropTypes>((props, ref) 
             fixed
             onTabSelect={onTabItemSelect}
             data-component-name="ObjectPageTabContainer"
-            className={classes.tabContainerComponent}
+            className={classNames.tabContainerComponent}
           >
             {safeGetChildrenArray(children).map((section, index) => {
               if (!isValidElement(section) || !section.props) return null;
@@ -917,16 +915,16 @@ const ObjectPage = forwardRef<HTMLDivElement, ObjectPagePropTypes>((props, ref) 
           </TabContainer>
         </div>
       )}
-      <div data-component-name="ObjectPageContent" className={classes.content} ref={objectPageContentRef}>
+      <div data-component-name="ObjectPageContent" className={classNames.content} ref={objectPageContentRef}>
         <div style={{ height: headerCollapsed ? `${headerContentHeight}px` : 0 }} aria-hidden />
         {placeholder ? placeholder : sections}
         <div style={{ height: `${sectionSpacer}px` }} aria-hidden />
       </div>
       {footer && mode === ObjectPageMode.IconTabBar && !sectionSpacer && (
-        <div className={classes.footerSpacer} data-component-name="ObjectPageFooterSpacer" aria-hidden />
+        <div className={classNames.footerSpacer} data-component-name="ObjectPageFooterSpacer" aria-hidden />
       )}
       {footer && (
-        <footer className={classes.footer} data-component-name="ObjectPageFooter">
+        <footer className={classNames.footer} data-component-name="ObjectPageFooter">
           {footer}
         </footer>
       )}
