@@ -1,6 +1,12 @@
 'use client';
 
-import { debounce, useI18nBundle, useIsomorphicLayoutEffect, useSyncRef } from '@ui5/webcomponents-react-base';
+import {
+  debounce,
+  useI18nBundle,
+  useIsomorphicLayoutEffect,
+  useStylesheet,
+  useSyncRef
+} from '@ui5/webcomponents-react-base';
 import { clsx } from 'clsx';
 import type { ElementType, HTMLAttributes, ReactElement, ReactNode, Ref, RefObject } from 'react';
 import React, {
@@ -14,7 +20,6 @@ import React, {
   useRef,
   useState
 } from 'react';
-import { createUseStyles } from 'react-jss';
 import type { PopupAccessibleRole } from '../../enums/index.js';
 import { ToolbarDesign, ToolbarStyle } from '../../enums/index.js';
 import { SHOW_MORE } from '../../i18n/i18n-defaults.js';
@@ -22,9 +27,7 @@ import { flattenFragments } from '../../internal/utils.js';
 import type { CommonProps } from '../../types/index.js';
 import type { ButtonPropTypes, PopoverDomRef, ToggleButtonPropTypes } from '../../webComponents/index.js';
 import { OverflowPopover } from './OverflowPopover.js';
-import { styles } from './Toolbar.jss.js';
-
-const useStyles = createUseStyles(styles, { name: 'Toolbar' });
+import { classNames, styleData } from './Toolbar.module.css.js';
 
 export interface ToolbarPropTypes extends Omit<CommonProps, 'onClick' | 'children'> {
   /**
@@ -166,7 +169,7 @@ const Toolbar = forwardRef<HTMLDivElement, ToolbarPropTypes>((props, ref) => {
     ...rest
   } = props;
 
-  const classes = useStyles();
+  useStylesheet(styleData, Toolbar.displayName);
   const [componentRef, outerContainer] = useSyncRef<HTMLDivElement>(ref);
   const controlMetaData = useRef([]);
   const [lastVisibleIndex, setLastVisibleIndex] = useState<number>(null);
@@ -180,12 +183,12 @@ const Toolbar = forwardRef<HTMLDivElement, ToolbarPropTypes>((props, ref) => {
   const showMoreText = i18nBundle.getText(SHOW_MORE);
 
   const toolbarClasses = clsx(
-    classes.outerContainer,
-    toolbarStyle === ToolbarStyle.Clear && classes.clear,
-    active && classes.active,
-    design === ToolbarDesign.Solid && classes.solid,
-    design === ToolbarDesign.Transparent && classes.transparent,
-    design === ToolbarDesign.Info && classes.info,
+    classNames.outerContainer,
+    toolbarStyle === ToolbarStyle.Clear && classNames.clear,
+    active && classNames.active,
+    design === ToolbarDesign.Solid && classNames.solid,
+    design === ToolbarDesign.Transparent && classNames.transparent,
+    design === ToolbarDesign.Info && classNames.info,
     className
   );
   const flatChildren = useMemo(() => {
@@ -207,12 +210,17 @@ const Toolbar = forwardRef<HTMLDivElement, ToolbarPropTypes>((props, ref) => {
         return item;
       }
       return (
-        <div ref={itemRef} key={index} className={classes.childContainer} data-component-name="ToolbarChildContainer">
+        <div
+          ref={itemRef}
+          key={index}
+          className={classNames.childContainer}
+          data-component-name="ToolbarChildContainer"
+        >
           {item}
         </div>
       );
     });
-  }, [flatChildren, controlMetaData, classes.childContainer]);
+  }, [flatChildren, controlMetaData, classNames.childContainer]);
 
   const overflowNeeded =
     (lastVisibleIndex || lastVisibleIndex === 0) &&
@@ -368,7 +376,7 @@ const Toolbar = forwardRef<HTMLDivElement, ToolbarPropTypes>((props, ref) => {
   return (
     <CustomTag
       style={styleWithMinWidth}
-      className={clsx(toolbarClasses, overflowNeeded && classes.hasOverflow)}
+      className={clsx(toolbarClasses, overflowNeeded && classNames.hasOverflow)}
       ref={componentRef}
       slot={slot}
       onClick={handleToolbarClick}
@@ -378,7 +386,7 @@ const Toolbar = forwardRef<HTMLDivElement, ToolbarPropTypes>((props, ref) => {
       data-sap-ui-fastnavgroup="true"
       {...rest}
     >
-      <div className={classes.toolbar} data-component-name="ToolbarContent" ref={contentRef}>
+      <div className={classNames.toolbar} data-component-name="ToolbarContent" ref={contentRef}>
         {overflowNeeded &&
           Children.map(childrenWithRef, (item, index) => {
             if (index >= lastVisibleIndex + 1 && index > numberOfAlwaysVisibleItems - 1) {
@@ -393,13 +401,13 @@ const Toolbar = forwardRef<HTMLDivElement, ToolbarPropTypes>((props, ref) => {
       {overflowNeeded && (
         <div
           ref={overflowBtnRef}
-          className={classes.overflowButtonContainer}
+          className={classNames.overflowButtonContainer}
           data-component-name="ToolbarOverflowButtonContainer"
         >
           <OverflowPopover
             overflowPopoverRef={overflowPopoverRef}
             lastVisibleIndex={lastVisibleIndex}
-            classes={classes}
+            classes={classNames}
             portalContainer={portalContainer}
             overflowContentRef={overflowContentRef}
             numberOfAlwaysVisibleItems={numberOfAlwaysVisibleItems}
