@@ -1,16 +1,9 @@
-import { isDesktop, isPhone, isTablet } from '@ui5/webcomponents-base/dist/Device.js';
 import searchIcon from '@ui5/webcomponents-icons/dist/search.js';
-import {
-  enrichEventWithDetails,
-  ThemingParameters,
-  useI18nBundle,
-  useIsomorphicId
-} from '@ui5/webcomponents-react-base';
+import { enrichEventWithDetails, useI18nBundle, useIsomorphicId, useStylesheet } from '@ui5/webcomponents-react-base';
 import type { MouseEventHandler, ReactNode } from 'react';
 import React, { Children, useEffect, useRef, useState } from 'react';
 import { createPortal } from 'react-dom';
-import { createUseStyles } from 'react-jss';
-import { BarDesign, FlexBoxAlignItems, FlexBoxDirection, ButtonDesign } from '../../enums/index.js';
+import { BarDesign, ButtonDesign, FlexBoxAlignItems, FlexBoxDirection } from '../../enums/index.js';
 import {
   APPLY_AUTOMATICALLY,
   CANCEL,
@@ -18,12 +11,11 @@ import {
   DEFAULT,
   MANAGE_VIEWS,
   SAVE,
+  SEARCH,
   SHARING,
-  VIEW,
-  SEARCH
+  VIEW
 } from '../../i18n/i18n-defaults.js';
 import { useCanRenderPortal } from '../../internal/ssr.js';
-import { cssVarVersionInfoPrefix } from '../../internal/utils.js';
 import { Bar } from '../../webComponents/Bar/index.js';
 import { Button } from '../../webComponents/Button/index.js';
 import { Dialog } from '../../webComponents/Dialog/index.js';
@@ -32,46 +24,10 @@ import { Icon, Input } from '../../webComponents/index.js';
 import { Table } from '../../webComponents/Table/index.js';
 import { TableColumn } from '../../webComponents/TableColumn/index.js';
 import { FlexBox } from '../FlexBox/index.js';
+import { classNames, styleData } from './ManageViewsDialog.module.css.js';
 import { ManageViewsTableRows } from './ManageViewsTableRows.js';
 import type { VariantManagementPropTypes } from './types.js';
 import type { VariantItemPropTypes } from './VariantItem.js';
-
-const _popupDefaultHeaderHeight = `var(${cssVarVersionInfoPrefix}popup_default_header_height)`;
-const _popupHeaderFontFamily = `var(${cssVarVersionInfoPrefix}popup_header_font_family)`;
-
-const styles = {
-  manageViewsDialog: {
-    // isTablet is true for some desktops with touch screens
-    width: isPhone() || (isTablet() && !isDesktop()) ? '100%' : '70vw',
-    '&::part(content), &::part(header)': {
-      padding: 0
-    },
-    '&::part(footer)': {
-      padding: 0,
-      borderBlockStart: 'none'
-    }
-  },
-  headerText: {
-    margin: 0,
-    textAlign: 'center',
-    alignSelf: 'start',
-    minHeight: _popupDefaultHeaderHeight,
-    maxHeight: _popupDefaultHeaderHeight,
-    lineHeight: _popupDefaultHeaderHeight,
-    textOverflow: 'ellipsis',
-    overflow: 'hidden',
-    whiteSpace: 'nowrap',
-    maxWidth: '100%',
-    display: 'inline-block',
-    paddingInlineStart: '1rem',
-    fontFamily: `"72override",${_popupHeaderFontFamily}`,
-    fontSize: '1rem'
-  },
-  search: { width: 'calc(100% - 2rem)', marginBlockEnd: '0.5rem' },
-  inputIcon: { cursor: 'pointer', color: ThemingParameters.sapContent_IconColor }
-};
-
-const useStyles = createUseStyles(styles, { name: 'ManageViewsDialog' });
 
 interface ManageViewsDialogPropTypes {
   children: ReactNode | ReactNode[];
@@ -123,7 +79,7 @@ export const ManageViewsDialog = (props: ManageViewsDialogPropTypes) => {
   const [changedVariantNames, setChangedVariantNames] = useState(new Map());
   const [invalidVariants, setInvalidVariants] = useState<Record<string, InputDomRef & { isInvalid?: boolean }>>({});
 
-  const classes = useStyles();
+  useStylesheet(styleData, 'ManageViewsDialog');
 
   const columns = (
     <>
@@ -258,7 +214,7 @@ export const ManageViewsDialog = (props: ManageViewsDialogPropTypes) => {
   return createPortal(
     <Dialog
       open
-      className={classes.manageViewsDialog}
+      className={classNames.manageViewsDialog}
       data-component-name="VariantManagementManageViewsDialog"
       onAfterClose={onAfterClose}
       onBeforeClose={handleClose}
@@ -266,13 +222,13 @@ export const ManageViewsDialog = (props: ManageViewsDialogPropTypes) => {
       initialFocus={`search-${uniqueId}`}
       header={
         <FlexBox direction={FlexBoxDirection.Column} style={{ width: '100%' }} alignItems={FlexBoxAlignItems.Center}>
-          <h2 className={classes.headerText}>{manageViewsText}</h2>
+          <h2 className={classNames.headerText}>{manageViewsText}</h2>
           <Input
             id={`search-${uniqueId}`}
-            className={classes.search}
+            className={classNames.search}
             placeholder={searchText}
             showClearIcon
-            icon={<Icon name={searchIcon} className={classes.inputIcon} />}
+            icon={<Icon name={searchIcon} className={classNames.inputIcon} />}
             onInput={handleSearchInput}
           />
         </FlexBox>
